Fix incorrect case in filename
