+++ src/main/java/com/aspose/barcode/cloud/requests/DeleteFileRequest.java
@@ -1,5 +1,5 @@
 // --------------------------------------------------------------------------------------------------------------------
-// <copyright company="Aspose" file="PutBarcodeRecognizeFromBodyRequest.java">
+// <copyright company="Aspose" file="DeleteFileRequest.java">
 //   Copyright (c) 2021 Aspose.BarCode for Cloud
 // </copyright>
 // <summary>
@@ -29,27 +29,17 @@
 
 package com.aspose.barcode.cloud.requests;
 
-import com.aspose.barcode.cloud.model.*;
+/** Delete file */
+public class DeleteFileRequest {
+    /** File path e.g. '/folder/file.ext' */
+    public final String path;
+    /** Storage name */
+    public String storageName;
+    /** File version ID to delete */
+    public String versionId;
 
-/** Recognition of a barcode from file on server with parameters in body. */
-public class PutBarcodeRecognizeFromBodyRequest {
-    /** The image file name. */
-    public final String name;
-    /** BarcodeReader object with parameters. */
-    public final ReaderParams readerParams;
-    /** */
-    public String type;
-    /** The storage name */
-    public String storage;
-    /** The image folder. */
-    public String folder;
-
-    /**
-     * @param name The image file name.
-     * @param readerParams BarcodeReader object with parameters.
-     */
-    public PutBarcodeRecognizeFromBodyRequest(String name, ReaderParams readerParams) {
-        this.name = name;
-        this.readerParams = readerParams;
+    /** @param path File path e.g. '/folder/file.ext' */
+    public DeleteFileRequest(String path) {
+        this.path = path;
     }
 }
