+++ src/main/java/com/aspose/barcode/cloud/requests/GetBarcodeGenerateRequest.java
@@ -1,5 +1,5 @@
 // --------------------------------------------------------------------------------------------------------------------
-// <copyright company="Aspose" file="PutBarcodeGenerateFileRequest.java">
+// <copyright company="Aspose" file="GetBarcodeGenerateRequest.java">
 //   Copyright (c) 2021 Aspose.BarCode for Cloud
 // </copyright>
 // <summary>
@@ -29,12 +29,8 @@
 
 package com.aspose.barcode.cloud.requests;
 
-import com.aspose.barcode.cloud.model.*;
-
-/** Generate barcode and save on server (from query params or from file with json or xml content) */
-public class PutBarcodeGenerateFileRequest {
-    /** The image file name. */
-    public final String name;
+/** Generate barcode. */
+public class GetBarcodeGenerateRequest {
     /** Type of barcode to generate. */
     public final String type;
     /** Text to encode. */
@@ -157,20 +153,14 @@ public class PutBarcodeGenerateFileRequest {
     public Double supplementSpace;
     /** Bars reduction value that is used to compensate ink spread while printing. */
     public Double barWidthReduction;
-    /** Image's storage. */
-    public String storage;
-    /** Image's folder. */
-    public String folder;
-    /** The image format. */
+    /** Result image format. */
     public String format;
 
     /**
-     * @param name The image file name.
      * @param type Type of barcode to generate.
      * @param text Text to encode.
      */
-    public PutBarcodeGenerateFileRequest(String name, String type, String text) {
-        this.name = name;
+    public GetBarcodeGenerateRequest(String type, String text) {
         this.type = type;
         this.text = text;
     }
