+++ src/main/java/com/aspose/barcode/cloud/requests/GetDiscUsageRequest.java
@@ -1,5 +1,5 @@
 // --------------------------------------------------------------------------------------------------------------------
-// <copyright company="Aspose" file="GetFilesListRequest.java">
+// <copyright company="Aspose" file="GetDiscUsageRequest.java">
 //   Copyright (c) 2021 Aspose.BarCode for Cloud
 // </copyright>
 // <summary>
@@ -29,17 +29,8 @@
 
 package com.aspose.barcode.cloud.requests;
 
-import com.aspose.barcode.cloud.model.*;
-
-/** Get all files and folders within a folder */
-public class GetFilesListRequest {
-    /** Folder path e.g. '/folder' */
-    public final String path;
+/** Get disc usage */
+public class GetDiscUsageRequest {
     /** Storage name */
     public String storageName;
-
-    /** @param path Folder path e.g. '/folder' */
-    public GetFilesListRequest(String path) {
-        this.path = path;
-    }
 }
