+++ src/main/java/com/aspose/barcode/cloud/requests/StorageExistsRequest.java
@@ -1,5 +1,5 @@
 // --------------------------------------------------------------------------------------------------------------------
-// <copyright company="Aspose" file="DownloadFileRequest.java">
+// <copyright company="Aspose" file="StorageExistsRequest.java">
 //   Copyright (c) 2021 Aspose.BarCode for Cloud
 // </copyright>
 // <summary>
@@ -29,19 +29,13 @@
 
 package com.aspose.barcode.cloud.requests;
 
-import com.aspose.barcode.cloud.model.*;
-
-/** Download file */
-public class DownloadFileRequest {
-    /** File path e.g. '/folder/file.ext' */
-    public final String path;
+/** Check if storage exists */
+public class StorageExistsRequest {
     /** Storage name */
-    public String storageName;
-    /** File version ID to download */
-    public String versionId;
+    public final String storageName;
 
-    /** @param path File path e.g. '/folder/file.ext' */
-    public DownloadFileRequest(String path) {
-        this.path = path;
+    /** @param storageName Storage name */
+    public StorageExistsRequest(String storageName) {
+        this.storageName = storageName;
     }
 }
